feat(chart): adjust the call structure of chart
